chore(rbac): promote nfs plugin from `./alpha` to primary entry point
