--- workspaces/rbac/plugins/rbac/dev/alpha/index.tsx
@@ -1,47 +1,0 @@
-/*
- * Copyright 2026 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * New Frontend System dev mode for the RBAC plugin.
- */
-// eslint-disable-next-line @backstage/no-ui-css-imports-in-non-frontend
-import '@backstage/ui/css/styles.css';
-import { createApp } from '@backstage/frontend-defaults';
-import ReactDOM from 'react-dom/client';
-
-import { createFrontendModule } from '@backstage/frontend-plugin-api';
-
-import rbacPlugin, { rbacTranslationsModule } from '../../src/alpha';
-import { devSidebarContent } from './shared';
-
-const devNavModule = createFrontendModule({
-  pluginId: 'app',
-  extensions: [devSidebarContent],
-});
-
-const defaultPage = '/rbac';
-
-const app = createApp({
-  features: [rbacPlugin, rbacTranslationsModule, devNavModule],
-});
-
-const root = app.createRoot();
-
-if (typeof window !== 'undefined' && window.location.pathname === '/') {
-  window.location.pathname = defaultPage;
-}
-
-ReactDOM.createRoot(document.getElementById('root')!).render(root);
