+++ workspaces/rbac/plugins/rbac/dev/shared.tsx
@@ -14,37 +14,134 @@
  * limitations under the License.
  */
 
-import { PropsWithChildren } from 'react';
-import Brightness2 from '@mui/icons-material/Brightness2';
-import WbSunny from '@mui/icons-material/WbSunny';
-
-import { AppTheme } from '@backstage/core-plugin-api';
 import {
-  UnifiedThemeProvider,
-  themes as builtinThemes,
-} from '@backstage/theme';
+  cloneElement,
+  ReactElement,
+  useCallback,
+  useState,
+  MouseEvent,
+} from 'react';
+import {
+  Sidebar,
+  SidebarGroup,
+  SidebarItem,
+  SidebarScrollWrapper,
+  SidebarSpace,
+} from '@backstage/core-components';
+import { appThemeApiRef, useApi } from '@backstage/core-plugin-api';
+import useObservable from 'react-use/esm/useObservable';
+import { NavContentBlueprint } from '@backstage/plugin-app-react';
+import ListItemIcon from '@mui/material/ListItemIcon';
+import ListItemText from '@mui/material/ListItemText';
+import Menu from '@mui/material/Menu';
+import MenuItem from '@mui/material/MenuItem';
+import AutoIcon from '@mui/icons-material/BrightnessAuto';
+import {
+  SidebarLanguageSwitcher,
+  SidebarSignOutButton,
+} from '@backstage/dev-utils';
 
-export const devAppThemes: AppTheme[] = [
-  {
-    id: 'light',
-    title: 'Light',
-    variant: 'light',
-    icon: <WbSunny />,
-    Provider: ({ children }: PropsWithChildren) => (
-      <UnifiedThemeProvider theme={builtinThemes.light}>
-        {children}
-      </UnifiedThemeProvider>
-    ),
+function ThemeIcon({
+  active,
+  icon,
+}: {
+  active?: boolean;
+  icon?: ReactElement;
+}) {
+  return icon ? (
+    cloneElement(icon, { color: active ? 'primary' : undefined })
+  ) : (
+    <AutoIcon color={active ? 'primary' : undefined} />
+  );
+}
+
+function SidebarThemeSwitcher() {
+  const appThemeApi = useApi(appThemeApiRef);
+  const themeId = useObservable(
+    appThemeApi.activeThemeId$(),
+    appThemeApi.getActiveThemeId(),
+  );
+  const themes = appThemeApi.getInstalledThemes();
+  const activeTheme = themes.find(t => t.id === themeId);
+  const [anchorEl, setAnchorEl] = useState<HTMLElement>();
+  const open = Boolean(anchorEl);
+
+  const handleOpen = (event: MouseEvent<Element>) => {
+    setAnchorEl(event.currentTarget as HTMLElement);
+  };
+  const handleSelectTheme = (newThemeId?: string) => {
+    if (newThemeId && themes.some(t => t.id === newThemeId)) {
+      appThemeApi.setActiveThemeId(newThemeId);
+    } else {
+      appThemeApi.setActiveThemeId(undefined);
+    }
+    setAnchorEl(undefined);
+  };
+  const handleClose = () => setAnchorEl(undefined);
+
+  const ActiveIcon = useCallback(
+    () => <ThemeIcon icon={activeTheme?.icon} />,
+    [activeTheme],
+  );
+
+  return (
+    <>
+      <SidebarItem icon={ActiveIcon} text="Switch Theme" onClick={handleOpen} />
+      <Menu
+        anchorEl={anchorEl}
+        open={open}
+        onClose={handleClose}
+        MenuListProps={{ role: 'listbox' }}
+      >
+        <MenuItem disabled>Choose a theme</MenuItem>
+        <MenuItem
+          selected={themeId === undefined}
+          onClick={() => handleSelectTheme(undefined)}
+        >
+          <ListItemIcon>
+            <ThemeIcon active={themeId === undefined} />
+          </ListItemIcon>
+          <ListItemText>Auto</ListItemText>
+        </MenuItem>
+        {themes.map(theme => {
+          const active = theme.id === themeId;
+          return (
+            <MenuItem
+              key={theme.id}
+              selected={active}
+              onClick={() => handleSelectTheme(theme.id)}
+            >
+              <ListItemIcon>
+                <ThemeIcon icon={theme.icon} active={active} />
+              </ListItemIcon>
+              <ListItemText>{theme.title}</ListItemText>
+            </MenuItem>
+          );
+        })}
+      </Menu>
+    </>
+  );
+}
+
+export const devSidebarContent = NavContentBlueprint.make({
+  params: {
+    component: ({ navItems }) => {
+      const nav = navItems.withComponent(item => (
+        <SidebarItem icon={() => item.icon} to={item.href} text={item.title} />
+      ));
+      return (
+        <Sidebar>
+          <SidebarGroup label="Menu">
+            <SidebarScrollWrapper>
+              {nav.rest({ sortBy: 'title' })}
+            </SidebarScrollWrapper>
+          </SidebarGroup>
+          <SidebarSpace />
+          <SidebarThemeSwitcher />
+          <SidebarLanguageSwitcher />
+          <SidebarSignOutButton />
+        </Sidebar>
+      );
+    },
   },
-  {
-    id: 'dark',
-    title: 'Dark',
-    variant: 'dark',
-    icon: <Brightness2 />,
-    Provider: ({ children }: PropsWithChildren) => (
-      <UnifiedThemeProvider theme={builtinThemes.dark}>
-        {children}
-      </UnifiedThemeProvider>
-    ),
-  },
-];
+});
