+++ workspaces/rbac/plugins/rbac/package.json
@@ -21,12 +21,16 @@
   "exports": {
     ".": "./src/index.ts",
     "./alpha": "./src/alpha/index.ts",
+    "./legacy": "./src/legacy/index.ts",
     "./package.json": "./package.json"
   },
   "typesVersions": {
     "*": {
       "alpha": [
         "src/alpha/index.ts"
+      ],
+      "legacy": [
+        "src/legacy/index.ts"
       ],
       "package.json": [
         "package.json"
@@ -41,15 +45,15 @@
     "postpack": "backstage-cli package postpack",
     "prepack": "backstage-cli package prepack",
     "start": "backstage-cli package start",
-    "start:alpha": "backstage-cli package start --entrypoint dev/alpha/index.tsx",
+    "start:legacy": "backstage-cli package start --entrypoint dev/legacy/index.tsx",
     "start:mock": "backstage-cli package start --entrypoint dev/index.mock.tsx",
-    "start:alpha:mock": "backstage-cli package start --entrypoint dev/alpha/index.mock.tsx",
+    "start:legacy:mock": "backstage-cli package start --entrypoint dev/legacy/index.mock.tsx",
     "test": "backstage-cli package test --passWithNoTests --coverage",
     "tsc": "tsc",
     "prettier:check": "prettier --ignore-unknown --check .",
     "prettier:fix": "prettier --ignore-unknown --write .",
-    "ui-test": "start-server-and-test start:mock localhost:3000 'APP_MODE=legacy playwright test -c ../../playwright.config.ts'",
-    "ui-test:alpha": "start-server-and-test start:alpha:mock localhost:3000 'APP_MODE=alpha playwright test -c ../../playwright.config.ts'"
+    "ui-test": "start-server-and-test start:mock localhost:3000 'playwright test -c ../../playwright.config.ts'",
+    "ui-test:legacy": "start-server-and-test start:legacy:mock localhost:3000 'APP_MODE=legacy playwright test -c ../../playwright.config.ts'"
   },
   "dependencies": {
     "@backstage-community/plugin-rbac-common": "workspace:^",
