+++ workspaces/rbac/plugins/rbac/src/legacy/index.ts
@@ -13,5 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+export { rbacPlugin, RbacPage, Administration } from '../plugin';
+export { rbacApiRef } from '../api/RBACBackendClient';
+export type { RBACAPI } from '../api/RBACBackendClient';
 
-export { rbacTranslationRef, rbacTranslations } from './translations';
+export { default as AdminPanelSettingsOutlinedIcon } from '@mui/icons-material/AdminPanelSettingsOutlined';
+export { default as RbacIcon } from '@mui/icons-material/VpnKeyOutlined';
+export type {
+  MemberEntity,
+  RoleError,
+  PluginConditionRules,
+  RoleBasedConditions,
+  ConditionRule,
+} from '../types';
